Add Category route CRUD
I added the code for the category CRUD in the routes/category.mjs files
where i implemented the standard rest endpoints to manage the category model.
This time i chose to return the category in case of success rather than a
message. In case of an error i just return the message.