Add dashboard links and category create
I creates the admin folder to hold the file that will manage the resources on the site.
i started with category because it is required by the product model and i will
create all the necessary routes for the category route before the product routes even
those that the client is supposed to see.