Add  client side category and product list
I added the pages to view all categories in a list and when one category is clicked
it lists all the products in that category. I also implemented the single product page.
I modeled it after the gulcan.com.tr website's single product page where you have the
swiper carousels with the product images and similar products.